% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellxgene3k}
\alias{cellxgene3k}
\title{Gene Counts From Cellxgene Census}
\format{
Seurat Sbject
}
\source{
CZ Cellxgene census, USA
}
\usage{
cellxgene3k
}
\description{
An RNAseq experiment subsampled from Cellxgene Census blood cell collection
with 3k cells.
}
\examples{
\dontrun{
 cellxgene3k
}
}
\references{
CZ CELLxGENE Discover: A single-cell data platform for scalable
exploration, analysis and modeling of aggregated data CZI Single-Cell
Biology, et al. bioRxiv 2023.10.30;
doi: https://doi.org/10.1101/2023.10.30.563174
}
\keyword{datasets}
