% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid.pbmc3k}
\alias{covid.pbmc3k}
\title{Gene Counts For COVID Patients}
\format{
Seurat Sbject
}
\source{
Chan Zuckerberg Initiative, USA
}
\usage{
covid.pbmc3k
}
\description{
An RNAseq experiment of Peripheral Blood Mononuclear Cells (PBMC) from
COVID-19 patients
}
\examples{
\dontrun{
 covid.pbmc3k
}
}
\references{
Ballestar, Esteban, et al.
"Single cell profiling of COVID-19 patients: an
international data resource from multiple tissues.” medRxiv (2020)."
URL https://www. medrxiv. org/content/10.1101/2020.11 20: v1.
}
\keyword{datasets}
