% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataInterface.R
\name{InstallDataset}
\alias{InstallDataset}
\title{Build the reference dataset}
\usage{
InstallDataset(datasetName = "cellxgene10k", cache_dir = "cache")
}
\arguments{
\item{datasetName}{A string of the dataset name}
}
\value{
Dataset Seurat object of the reference dataset
}
\description{
A function to build the reference dataset given the dataset name at cache directory
}
\details{
The supported dataset names are:
\itemize{
\item "cellxgene10k": the default dataset
}
}
\examples{
# Example1: Install the default dataset (cellxgene10k)
InstallDataset("cellxgene10k")

}
