% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotAnnotation}
\alias{plotAnnotation}
\title{Plot Annotation}
\usage{
plotAnnotation(queryDataset, prediction)
}
\arguments{
\item{queryDataset}{query dataset}

\item{prediction}{reference mapping result}
}
\value{
a figure of annotation
}
\description{
A function to plot the annotation of query dataset using UMAP
dimensionality reduction with Seurat.
}
\examples{
predictions <- searchCell(covid.pbmc3k, cellxgene3k)
plotAnnotation(covid.pbmc3k, predictions)

}
\references{
McInnes, Leland,
John Healy, and James Melville. "Umap: Uniform manifold approximation
and projection for dimension reduction."
arXiv preprint arXiv:1802.03426 (2018).
}
