% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceMapping.R
\name{referenceMappingSeurat}
\alias{referenceMappingSeurat}
\title{Perform reference mapping with Seurat}
\usage{
referenceMappingSeurat(queryDataset, referenceDataset)
}
\arguments{
\item{queryDataset}{query dataset}

\item{referenceDataset}{reference dataset}
}
\value{
predictions of reference mapping
}
\description{
A function to perform reference mapping with Seurat.
}
