% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCellSearchR.R
\name{runCellSearchR}
\alias{runCellSearchR}
\title{Launch Shiny App for CellSearchR}
\usage{
runCellSearchR()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for CellSearchR
This app allows users to upload a query dataset and choose a reference
datset for reference mapping.
}
\examples{
\dontrun{

CellSearchR::runCellSearchR()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
