% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceMapping.R
\name{searchCell}
\alias{searchCell}
\title{Perform reference mapping}
\usage{
searchCell(queryDataset, referenceDataset, embedMethod = "Seurat")
}
\arguments{
\item{queryDataset}{A Seurat object of the query dataset}

\item{referenceDataset}{A Seurat object of the reference dataset;}

\item{embedMethod}{A string of the embedding method,
which is one of the following:
\itemize{
\item "Seurat": the default embedding method using Seurat
}}
}
\value{
predictions of reference mapping
}
\description{
A function to perform reference mapping, which is to map the query dataset to
the reference dataset with specified embedding method.
}
\examples{
predictions <- searchCell(covid.pbmc3k, cellxgene3k,
                           embedMethod = "Seurat")

}
\references{
Hao Y, Hao S, Andersen-Nissen E, III WMM, Zheng S, Butler A, Lee MJ, Wilk AJ,
Darby C, Zagar M, Hoffman P, Stoeckius M, Papalexi E, Mimitou EP, Jain J,
Srivastava A, Stuart T, Fleming LB, Yeung B, Rogers AJ, McElrath JM,
Blish CA, Gottardo R, Smibert P, Satija R (2021). “Integrated analysis of
multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048,
https://doi.org/10.1016/j.cell.2021.04.048.
}
