% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Embedder.R
\name{seuratEmbedding}
\alias{seuratEmbedding}
\title{Generate Seurat embeddings for both reference and query dataset}
\usage{
seuratEmbedding(referenceDataset, queryDataset)
}
\arguments{
\item{referenceDataset}{preprocessed reference dataset}

\item{queryDataset}{preprocessed query dataset}
}
\value{
Seurat embeddings
}
\description{
A utility function to generate Seurat embeddings
for both reference and query dataset.
}
\details{
Seurat embeddings uses PCA-based approach across both reference and query
dataset.
}
